feat!: block center squares for balancing
The changes actually gives earlier players, especially P1, a big
advantage since they can start in the center (also it's less fun since
it's an obvious power move). Since the center is so powerful, I decided
to not allow anyone to start there. This makes the game much more
interesting and can actually be advantageous to both earlier and later
players depending on how it plays out.